Micro_Services Version/service_a_perception/requirements.txt
faster-whisper
pyannote.audio
numpy<2.0
transformers==4.40.0 
huggingface-hub==0.22.2
demucs
fastapi
uvicorn
pydantic